tensorflow/minigo_0_6/tensorflow/minigo/requirements.txt
absl-py
autopep8>=1.3
fire
google.cloud.logging
google.cloud.bigtable
grpcio-tools
keras
numpy>=1.14.0
protobuf
pylint
sgf==0.5
six
tqdm>=4.17
pyasn1>=0.4.1
setuptools>=34.0.0

oauth2client==4.1

gsutil

pytest==5.4.1
pytest-pythonpath==0.7.3
pytest-mock==2.0.0